Creating two seperate modules one for runtime and one for the editor. This will allow us to start supporting runtime solutions encased in the same plugin
