--- Source/Mythica/Public/Mythica.h
@@ -1,20 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
-
-class FMythicaModule : public IModuleInterface
-{
-public:
-
-    /** IModuleInterface implementation */
-    virtual void StartupModule() override;
-    virtual void ShutdownModule() override;
-
-    void OpenPackageManager();
-
-private:
-    void AddMenu(FMenuBuilder& MenuBuilder);
-};
