+++ Source/MythicaEditor/Public/MythicaEditor.h
@@ -1,0 +1,19 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "Modules/ModuleManager.h"
+
+class FMythicaEditorModule : public IModuleInterface
+{
+public:
+
+    /** IModuleInterface implementation */
+    virtual void StartupModule() override;
+    virtual void ShutdownModule() override;
+
+    void OpenPackageManager();
+
+private:
+    void AddMenu(FMenuBuilder& MenuBuilder);
+};
